create room folder, add set skybox message to it
