+++ nanome/api/room/messages.py
@@ -1,0 +1,19 @@
+from nanome._internal import serializer_fields
+
+
+class SetSkybox(serializer_fields.TypeSerializer):
+
+    def __init__(self):
+        pass
+
+    def version(self):
+        return 0
+
+    def name(self):
+        return "SetSkybox"
+
+    def serialize(self, version, value, context):
+        context.write_int(value)
+
+    def deserialize(self, version, context):
+        raise NotImplementedError
